{"schema":"decodex/commit/1","summary":"Modularize settings surface","authority":"manual"}
